# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://gabomapay.com/api"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Authentifie l'utilisateur avant de se connecter",
                "description": "",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Deconnexion de l'utilisateur",
                "description": "",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "schema": "Bearer",
                                        "type": "string",
                                        "example": "89er4186gjazuihhiZIOJreioiouEIOJIOZERF814879AE8FEA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Inscription de l'utilisateur",
                "description": "",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "country": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+243852277378"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/forgot": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Demande de rénitialisation de mot de passe par mail",
                "description": "",
                "operationId": "03aca08208571c2a43bcdeebb40304ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/reset": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Renitialisation du mot de passe",
                "description": "",
                "operationId": "8ae8881c540d1f799293343fde43c530",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "ufghzruqgioqzruogrjcuuUIBYRCUR?61154600"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/profile": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Modification des informations du profile de l'utilisateur",
                "description": "",
                "operationId": "c9645f8c93e992e4df8b013e542bd367",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Kabamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+243852277378"
                                    },
                                    "sexe": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "N°5 Bukasa Limete 1iere Rue"
                                    },
                                    "profile": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/identite": {
            "post": {
                "tags": [
                    "AUHTENTIFICATION"
                ],
                "summary": "Chargement de la piéce d'identité de l'utilisateur",
                "description": "",
                "operationId": "c67945b64410b2980d9d6b9c6cf9e969",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "identite": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}